chore: drawer performance enhancements
